requirements.txt
django==5.0.1
numpy==1.26.3
pandas==2.2.0
pyarrow==14.0.1
scikit-learn==1.4.0
tensorflow==2.16.1
django-bootstrap5==23.4
plotly==5.0.0
python-dotenv==1.0.0
joblib==1.3.2
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.0
textblob==0.17.1
scipy==1.12.0